# contracts/openapi.v0.1.yaml
openapi: 3.1.0
info:
  title: DSH 工业仿真智能体 工程 API
  version: 0.1.2
  description: |
    dsh-sim R0 工程能力真相入口（定义书 §工程API：业务操作 / §工程API：审查、节点与上传）。
    MCP 是本 API 的受限投影。长操作返回 202 与资源 ID；请求返回不是计算结束。
    所有产生副作用的 POST 必须携带 Idempotency-Key 头（主体+项目+action+key 唯一；
    同 key 同摘要返回原对象，同 key 异摘要 → 409）。
    错误对象统一为 Error{code,message,retryable,trace_id,details}。
    本文件中的示例值均为占位：假哈希一律写作 "sha256:<64hex>" 形式，不含真实工程参数。
servers:
  - url: http://127.0.0.1:8600/api/v1
...
paths:
...
  /tasks:
    post:
      tags: [tasks]
      operationId: createTask
      summary: 创建任务草稿（符合 task-draft.schema.json）
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [draft]
              properties:
                draft:
                  type: object
                  description: 符合 https://dsh-sim.local/contracts/task-draft.schema.json
            example:
              draft:
                purpose: design_screening
                open_questions:
                  - field: conditions
                    responsible: "<engineer_id>"
                    question: "<missing input question>"
      responses:
        '201':
          description: 已创建任务
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /bundles/{bundle_id}:
    get:
      tags: [bundles]
      operationId: getBundle
      summary: 读取冻结证据包清单与摘要
      parameters:
        - name: bundle_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Bundle 对象
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Bundle' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /tasks/{task_id}:
    get:
      tags: [projections]
      operationId: getTaskProjection
      summary: 读取任务投影（含 evidence_mode 与归一 blockers）
      x-extension: read-projection
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task 投影
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Task'
                  - type: object
                    properties:
                      evidence_mode: { enum: [REAL, MOCK, UNKNOWN] }
                      blocker_objects:
                        type: array
                        items: { $ref: '#/components/schemas/Blocker' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /tasks/{task_id}/revisions/latest:
    get:
      tags: [projections]
      operationId: getLatestRevisionProjection
      summary: 当前修订投影
      x-extension: read-projection
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 当前 TaskRevision
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskRevision' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /tasks/{task_id}/runs:
    get:
      tags: [projections]
      operationId: listRunProjection
      summary: 矩阵投影 {items}——每 run 带 variant/condition/三维状态/事件与最新阶段
      x-extension: read-projection
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Run 投影列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/Run'
                        - type: object
                          properties:
                            latest_stage: { type: string, nullable: true }
                            latest_event_at: { type: string, format: date-time, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /tasks/{task_id}/events:
    get:
      tags: [projections]
      operationId: listTaskEventsIncremental
      summary: 跨 job 增量事件流——(occurred_at,event_seq) 全局排序重编号 global_seq，after_seq 增量拉取
      x-extension: read-projection
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - name: after_seq
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 0 }
          description: 上次响应的 next_after（global_seq 游标）；只返回其后的新事件。
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 200 }
      responses:
        '200':
          description: 增量事件列表
          content:
            application/json:
              schema:
                type: object
                required: [items, next_after]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [global_seq, job_id, event_seq, kind, payload, occurred_at]
                      properties:
                        global_seq: { type: integer, minimum: 1 }
                        job_id: { type: string }
                        event_seq: { type: integer, minimum: 1 }
                        kind: { type: string }
                        payload: { type: object }
                        occurred_at: { type: string, format: date-time }
                  next_after:
                    type: integer
                    minimum: 0
                    description: 本次返回的最大 global_seq；无新事件时等于请求的 after_seq。
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /tasks/{task_id}/preparations/latest:
    get:
      tags: [projections]
      operationId: getLatestPreparationProjection
      summary: 最新准备投影——differences[].field/requested/readback/unit/boundary/status + blockers
      x-extension: read-projection
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Preparation 投影
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Preparation'
                  - type: object
                    properties:
                      ready: { type: boolean }
                      evidence_mode: { enum: [REAL, MOCK, UNKNOWN] }
                      blocker_objects:
                        type: array
                        items: { $ref: '#/components/schemas/Blocker' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /tasks/{task_id}/bundles/latest:
    get:
      tags: [projections]
      operationId: getLatestBundleProjection
      summary: 最新证据包投影（含 evidence_mode/complete/incomplete_items/uncovered_scope）
      x-extension: read-projection
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Bundle 投影
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Bundle'
                  - type: object
                    properties:
                      evidence_mode: { enum: [REAL, MOCK, UNKNOWN] }
                      complete: { type: boolean }
                      incomplete_items:
                        type: array
                        items: { type: object }
                      uncovered_scope: { type: string }
                      limitations: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /tasks/{task_id}/review-history:
    get:
      tags: [projections]
      operationId: getReviewHistoryProjection
      summary: 跨修订审查/决定历史（"当时接受 R1"与"当前 R2 尚未接受"两态可读，FR-23）
      x-extension: read-projection
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 审查历史
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /bundles/{bundle_id}/manifest:
    get:
      tags: [projections]
      operationId: getBundleManifestProjection
      summary: 全量冻结清单投影（含 role/length/complete/incomplete_items/uncovered_scope）
      x-extension: read-projection
      parameters:
        - name: bundle_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 证据包投影（manifest 条目含 artifact_id/logical_path/sha256/length/role）
          content:
            application/json:
              schema:
                type: object
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /bundles/{bundle_id}/claims:
    get:
      tags: [projections]
      operationId: listBundleClaims
      summary: 包内 Claim 列表（每条绑定 metric_id + artifact_id）
      x-extension: read-projection
      parameters:
        - name: bundle_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Claim 列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Claim' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /bundles/{bundle_id}/verifications:
    get:
      tags: [projections]
      operationId: listBundleVerifications
      summary: 包内 Verification 记录
      x-extension: read-projection
      parameters:
        - name: bundle_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Verification 列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Verification' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
      description: 幂等键；主体+项目+action+key 唯一。同 key 同请求摘要返回原对象；同 key 异摘要 → 409。
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema: { type: string }
  responses:
    Unauthorized:
      description: 401 未认证
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: 403 越权
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: 409 摘要/修订/幂等冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unprocessable:
      description: 422 工程输入不成立
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooManyRequests:
      description: 429 配额限制
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ServiceUnavailable:
      description: 503 暂不可用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Blocker:
      type: object
      required: [code, message, responsible, action]
      description: 阻塞项统一形状（0.1.1 读取投影）：阻塞不是静默跳过
      properties:
        code: { type: string }
        message: { type: string }
        responsible: { type: string }
        action: { type: string }
    TaskFlowState:
      enum: [DRAFT, PREPARING, READY, AUTHORIZED, ACTIVE, READY_FOR_REVIEW, IN_REVIEW, ACCEPTED, CHANGES_REQUESTED, REJECTED]
...
    Task:
      type: object
      required: [task_id, project_id, current_revision, task_state, review_state, owner_id, purpose, created_at]
      properties:
        task_id: { type: string, description: 稳定任务 ID }
        project_id: { type: string }
        current_revision: { type: integer, minimum: 0 }
        task_state: { $ref: '#/components/schemas/TaskFlowState' }
        review_state: { $ref: '#/components/schemas/ReviewState' }
        owner_id: { type: string }
        purpose: { type: string, enum: [design_screening] }
        blockers:
          type: array
          items: { type: object }
          description: 阻塞不是静默跳过；缺输入/回读失败在此表达
        created_at: { type: string, format: date-time }
    TaskRevision:
      type: object
      required: [task_id, revision, spec, spec_sha256, source_refs, created_by, created_at]
      properties:
        task_id: { type: string }
        revision: { type: integer, minimum: 1, description: 单调递增；发布后不可原位修改 }
        spec:
          type: object
          description: 完整业务输入，符合 task-spec.schema.json
        spec_sha256: { $ref: '#/components/schemas/Sha256' }
        source_refs:
          type: array
          items: { type: string }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
...
    Run:
      type: object
      required: [run_id, task_id, revision, variant_id, condition_id, execution_state, numerical_state, applicability_state, created_at]
      properties:
        run_id: { type: string }
        task_id: { type: string }
        revision: { type: integer, minimum: 1 }
        variant_id: { type: string }
        condition_id: { type: string }
        execution_state: { $ref: '#/components/schemas/ExecutionState' }
        numerical_state: { $ref: '#/components/schemas/NumericalState' }
        applicability_state: { $ref: '#/components/schemas/ApplicabilityState' }
        current_attempt_id: { type: string, nullable: true }
        events:
          type: array
          items: { $ref: '#/components/schemas/Event' }
          description: 按 after_seq 增量返回
        created_at: { type: string, format: date-time }
...
    Bundle:
      type: object
      required: [bundle_id, task_id, revision, bundle_digest, manifest, validity, created_at]
      properties:
        bundle_id: { type: string }
        task_id: { type: string }
        revision: { type: integer, minimum: 1 }
        bundle_digest: { $ref: '#/components/schemas/Sha256' }
        manifest:
          type: array
          items:
            type: object
            required: [artifact_id, logical_path, sha256]
            properties:
              artifact_id: { type: string }
              logical_path: { type: string }
              sha256: { $ref: '#/components/schemas/Sha256' }
          description: 冻结清单，不可原位改写
        validity: { $ref: '#/components/schemas/ValidityState' }
        created_at: { type: string, format: date-time }
...